Finished boot  functionality
