+++ index.html
@@ -146,17 +146,17 @@
     <button class="trigger-event btn btn-secondary" rel="year">Trigger year change</button>
 </div>
 
+<div id="game-area">
+    <div id="boot-container"></div>
 
-<div id="boots">
-    <template id="boot">
-        Amount of tickets sold: <span class="boot-ticket-sold"></span>
-        Amount of customers waiting: <span class="boot-customer-waiting"></span>
-        Amount of customers who went away: <span class="boot-customer-away"></span>
+    <template id="boot-template">
+        <p>Amount of tickets sold: <span class="boot-ticket-sold"></span></p>
+        <p>Amount of customers waiting: <span class="boot-customer-waiting"></span></p>
+        <p>Amount of customers who went away: <span class="boot-customer-away"></span></p>
     </template>
 
     <button class="btn btn-secondary" id="build-boot">Build boot</button>
 </div>
-
 
 </body>
 </html>
